# 9. Sprint III/pf-be-hisp-w26-t07-torres/docs/specs/swagger.yaml
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "pf-be-hisp-w26-t07-torres",
    "description" : "pf-be-hisp-w26-t07-torres",
    "contact" : {
      "name" : "carcardozo",
      "email" : "carlos.cardozo@mercadolibre.com.co"
    },
    "version" : "last"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Scope local"
  } ],
  "paths" : {
    "/api/v1/fresh-products/{idSeller}/{idProduct}" : {
      "put" : {
        "tags" : [ "product-controller" ],
        "operationId" : "updateProductSeller",
        "parameters" : [ {
          "name" : "idProduct",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "idSeller",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProductRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductCreateResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/fresh-products/{idSeller}" : {
      "post" : {
        "tags" : [ "product-controller" ],
        "operationId" : "createProductSeller",
        "parameters" : [ {
          "name" : "idSeller",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ListProductRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductCreateResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/authenticate/seller/{id}" : {
      "get" : {
        "tags" : [ "authenticate-controller" ],
        "operationId" : "authenticateSeller",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TokenResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/authenticate/representative/{id}" : {
      "get" : {
        "tags" : [ "authenticate-controller" ],
        "operationId" : "authenticateRepresentative",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TokenResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/authenticate/buyer/{id}" : {
      "get" : {
        "tags" : [ "authenticate-controller" ],
        "operationId" : "authenticateBuyer",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TokenResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fresh-products/{idProduct}/warehouse/list" : {
      "get" : {
        "tags" : [ "product-controller" ],
        "operationId" : "getStockQuantityForEachWarehouse",
        "parameters" : [ {
          "name" : "idProduct",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StockQuantityResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fresh-products/{idProduct}/batch/list" : {
      "get" : {
        "tags" : [ "product-controller" ],
        "operationId" : "getBatchListByProduct",
        "parameters" : [ {
          "name" : "idProduct",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "order",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LocationForProductDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fresh-products/list" : {
      "get" : {
        "tags" : [ "product-controller" ],
        "operationId" : "getAllProducts",
        "parameters" : [ {
          "name" : "category",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ProductResponseDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "ProductRequestDto" : {
        "required" : [ "product_desc", "storage_type" ],
        "type" : "object",
        "properties" : {
          "price" : {
            "type" : "number",
            "format" : "double"
          },
          "product_desc" : {
            "type" : "string"
          },
          "storage_type" : {
            "type" : "string"
          }
        }
      },
      "ProductCreateResponseDto" : {
        "type" : "object",
        "properties" : {
          "operation" : {
            "type" : "integer",
            "format" : "int64"
          },
          "message" : {
            "type" : "string"
          },
          "code" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "ListProductRequestDto" : {
        "required" : [ "products" ],
        "type" : "object",
        "properties" : {
          "products" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ProductRequestDto"
            }
          }
        }
      },
      "TokenResponseDto" : {
        "type" : "object",
        "properties" : {
          "token" : {
            "type" : "string"
          }
        }
      },
      "StockQuantityResponseDto" : {
        "type" : "object",
        "properties" : {
          "product_id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "warehouses" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/WarehouseResponseDTO"
            }
          }
        }
      },
      "WarehouseResponseDTO" : {
        "type" : "object",
        "properties" : {
          "warehouse_code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "total_quantity" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "BSResponseDTO" : {
        "type" : "object",
        "properties" : {
          "batch_number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "current_quantity" : {
            "type" : "integer",
            "format" : "int32"
          },
          "due_date" : {
            "type" : "string",
            "format" : "date"
          }
        }
      },
      "LocationForProductDTO" : {
        "type" : "object",
        "properties" : {
          "section" : {
            "$ref" : "#/components/schemas/SectionResponseDTO"
          },
          "product_id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "batch_stock" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/BSResponseDTO"
            }
          }
        }
      },
      "SectionResponseDTO" : {
        "type" : "object",
        "properties" : {
          "section_code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "warehouse_code" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "ProductResponseDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "description" : {
            "type" : "string"
          },
          "price" : {
            "type" : "number"
          },
          "seller_name" : {
            "type" : "string"
          },
          "category" : {
            "type" : "string"
          }
        }
      },
...
    }
  }
}